+++ crates/mimic/src/design/base/sanitizer/mod.rs
@@ -1,4 +1,5 @@
 pub mod intl;
 pub mod num;
 pub mod text;
+pub mod time;
 pub mod web;