+++ crates/mimic_declare/src/node/entity.rs
@@ -73,12 +73,7 @@ impl HasSchemaPart for Entity {
 impl HasTraits for Entity {
     fn traits(&self) -> TraitList {
         let mut traits = self.traits.clone().with_type_traits();
-        traits.extend(vec![
-            Trait::Inherent,
-            Trait::EntityKind,
-            Trait::EntityLifecycle,
-            Trait::FieldValues,
-        ]);
+        traits.extend(vec![Trait::Inherent, Trait::EntityKind, Trait::FieldValues]);
 
         traits.list()
     }
@@ -92,7 +87,6 @@ impl HasTraits for Entity {
             Trait::Default => DefaultTrait::strategy(self),
             Trait::From => FromTrait::strategy(self),
             Trait::EntityKind => EntityKindTrait::strategy(self),
-            Trait::EntityLifecycle => EntityLifecycleTrait::strategy(self),
             Trait::FieldValues => FieldValuesTrait::strategy(self),
             Trait::TypeView => TypeViewTrait::strategy(self),
             Trait::SanitizeAuto => SanitizeAutoTrait::strategy(self),