+++ crates/mimic_declare/src/schema_traits.rs
@@ -64,7 +64,6 @@ pub enum Trait {
     StoreKind,
 
     // orm
-    EntityLifecycle,
     FieldValue,
     FieldValues,
     From,