+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable, and occasionally less notable changes to this project will be docum
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.24.29] - 2025-10-21
+- removed the EntityLifecycle traits, and now we just have a simple CreatedAt/UpdatedAt
+that mutates via the sanitizer that can be used multiple places
+
 ## [0.24.21] - 2025-10-19
 - added a primitive Account type as we need to start using it in indexes
 - also added Account as a potential Key type